+++ suppressible-error-prone/src/main/java/com/palantir/suppressibleerrorprone/VisitorStateModifications.java
@@ -18,40 +18,21 @@ package com.palantir.suppressibleerrorprone;
 
 // CHECKSTYLE:OFF
 
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Range;
 import com.google.errorprone.VisitorState;
-import com.google.errorprone.fixes.Fix;
-import com.google.errorprone.fixes.Replacement;
-import com.google.errorprone.fixes.Replacements.CoalescePolicy;
-import com.google.errorprone.fixes.SuggestedFix;
 import com.google.errorprone.matchers.Description;
 import com.sun.source.tree.AnnotationTree;
-import com.sun.source.tree.AssignmentTree;
 import com.sun.source.tree.ClassTree;
-import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.IdentifierTree;
-import com.sun.source.tree.LiteralTree;
 import com.sun.source.tree.MemberSelectTree;
 import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.ModifiersTree;
-import com.sun.source.tree.NewArrayTree;
 import com.sun.source.tree.Tree;
 import com.sun.source.tree.VariableTree;
 import com.sun.source.util.TreePath;
-import com.sun.tools.javac.tree.EndPosTable;
-import com.sun.tools.javac.tree.JCTree.JCCompilationUnit;
-import com.sun.tools.javac.util.JCDiagnostic.DiagnosticPosition;
-import java.util.LinkedHashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.WeakHashMap;
-import java.util.function.Predicate;
-import java.util.function.Supplier;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.lang.model.element.Name;
 // CHECKSTYLE:ON
@@ -59,7 +40,7 @@ import javax.lang.model.element.Name;
 public final class VisitorStateModifications {
     private static final Logger log = Logger.getLogger(VisitorStateModifications.class.getName());
 
-    private static final Map<Tree, OurFix> FIXES = new WeakHashMap<>();
+    private static final Map<Tree, SuppressFix> FIXES = new WeakHashMap<>();
 
     @SuppressWarnings("RestrictedApi")
     public static Description interceptDescription(VisitorState visitorState, Description description) {
@@ -103,12 +84,12 @@ public final class VisitorStateModifications {
 
         boolean containedKey = FIXES.containsKey(firstSuppressibleParent);
 
-        OurFix ourFix = FIXES.computeIfAbsent(
+        SuppressFix suppressFix = FIXES.computeIfAbsent(
                 firstSuppressibleParent,
-                _ignored -> new OurFix(
+                _ignored -> new SuppressFix(
                         Optional.ofNullable(visitorState.getSourceCode()), suppressWarnings, firstSuppressibleParent));
 
-        ourFix.addError(description.checkName);
+        suppressFix.addError(description.checkName);
 
         if (containedKey) {
             return Description.NO_MATCH;
@@ -119,141 +100,8 @@ public final class VisitorStateModifications {
                         description.checkName,
                         description.getLink(),
                         description.getMessageWithoutCheckName())
-                .addFix(ourFix)
+                .addFix(suppressFix)
                 .build();
-    }
-
-    private static final class OurReplacement extends Replacement {
-        private final Supplier<Replacement> replacement;
-
-        OurReplacement(Supplier<Replacement> replacement) {
-            this.replacement = replacement;
-        }
-
-        @Override
-        public Range<Integer> range() {
-            return replacement.get().range();
-        }
-
-        @Override
-        public String replaceWith() {
-            return replacement.get().replaceWith();
-        }
-    }
-
-    private static final class OurFix implements Fix {
-        private final Optional<CharSequence> sourceCode;
-        private final Optional<? extends AnnotationTree> suppressWarnings;
-        private final Tree tree;
-        private final Set<String> errors = new LinkedHashSet<>();
-
-        OurFix(Optional<CharSequence> sourceCode, Optional<? extends AnnotationTree> suppressWarnings, Tree tree) {
-            this.sourceCode = sourceCode;
-            this.suppressWarnings = suppressWarnings;
-            this.tree = tree;
-        }
-
-        private static String suppressWarningsString(List<String> warningsToSuppress) {
-            String suppressWarningsString = '"' + String.join("\", \"", warningsToSuppress) + '"';
-
-            if (warningsToSuppress.size() > 1) {
-                suppressWarningsString = "{" + suppressWarningsString + "}";
-            }
-            return "@SuppressWarnings(" + suppressWarningsString + ")";
-        }
-
-        private CharSequence indentForTree() {
-            return sourceCode
-                    .map(actualSourceCode ->
-                            whitespaceIndentBefore(actualSourceCode, ((DiagnosticPosition) tree).getStartPosition()))
-                    .orElse("    ");
-        }
-
-        private Fix fix() {
-            return suppressWarnings
-                    .map(suppressWarningsAnnotation -> {
-                        Set<String> existingValues = annotationStringValues(suppressWarningsAnnotation)
-                                .collect(Collectors.toSet());
-                        Set<String> toAdd = errors.stream()
-                                .filter(Predicate.not(error -> existingValues.contains(error)
-                                        || existingValues.contains(CommonConstants.AUTOMATICALLY_ADDED_PREFIX + error)))
-                                .collect(Collectors.toSet());
-
-                        List<String> warningsToSuppress = Stream.concat(
-                                        existingValues.stream(),
-                                        toAdd.stream()
-                                                .sorted()
-                                                .map(warning -> CommonConstants.AUTOMATICALLY_ADDED_PREFIX + warning))
-                                .collect(Collectors.toList());
-
-                        String suppressWarningsString = suppressWarningsString(warningsToSuppress);
-
-                        return SuggestedFix.replace(suppressWarningsAnnotation, suppressWarningsString);
-                    })
-                    .orElseGet(() -> {
-                        List<String> warningsToSuppress = errors.stream()
-                                .sorted()
-                                .map(warning -> CommonConstants.AUTOMATICALLY_ADDED_PREFIX + warning)
-                                .collect(Collectors.toList());
-
-                        String suppressWarningsString = suppressWarningsString(warningsToSuppress);
-
-                        return SuggestedFix.prefixWith(tree, suppressWarningsString + "\n" + indentForTree());
-                    });
-        }
-
-        public void addError(String error) {
-            errors.add(error);
-        }
-
-        @Override
-        public String toString(JCCompilationUnit compilationUnit) {
-            return fix().toString(compilationUnit);
-        }
-
-        @Override
-        public String getShortDescription() {
-            return fix().getShortDescription();
-        }
-
-        @Override
-        public CoalescePolicy getCoalescePolicy() {
-            return CoalescePolicy.REJECT;
-        }
-
-        @Override
-        public ImmutableSet<Replacement> getReplacements(EndPosTable endPositions) {
-            return ImmutableSet.of(new OurReplacement(
-                    () -> fix().getReplacements(endPositions).iterator().next()));
-        }
-
-        @Override
-        public ImmutableSet<String> getImportsToAdd() {
-            return ImmutableSet.of();
-        }
-
-        @Override
-        public ImmutableSet<String> getImportsToRemove() {
-            return ImmutableSet.of();
-        }
-
-        @Override
-        public boolean isEmpty() {
-            return false;
-        }
-    }
-
-    static CharSequence whitespaceIndentBefore(CharSequence sourceCode, int sourceElementPosition) {
-        int pos = sourceElementPosition - 1;
-
-        for (; pos >= 0; pos--) {
-            char character = sourceCode.charAt(pos);
-            if (character == '\n' || !Character.isWhitespace(character)) {
-                break;
-            }
-        }
-
-        return sourceCode.subSequence(pos + 1, sourceElementPosition);
     }
 
     private static boolean suppressibleTree(Tree tree) {
@@ -289,33 +137,6 @@ public final class VisitorStateModifications {
 
         throw new UnsupportedOperationException(
                 "Unsupported annotation type: " + annotationType.getClass().getCanonicalName());
-    }
-
-    static Stream<String> annotationStringValues(AnnotationTree annotation) {
-        return annotation.getArguments().stream().flatMap(arg -> {
-            if (!(arg instanceof AssignmentTree)) {
-                return Stream.empty();
-            }
-            AssignmentTree assignment = (AssignmentTree) arg;
-
-            ExpressionTree expression = assignment.getExpression();
-
-            if (expression instanceof LiteralTree) {
-                LiteralTree literalTree = (LiteralTree) expression;
-                return Stream.of((String) literalTree.getValue());
-            }
-
-            if (expression instanceof NewArrayTree) {
-                NewArrayTree newArray = (NewArrayTree) expression;
-                return newArray.getInitializers().stream()
-                        .map(LiteralTree.class::cast)
-                        .map(LiteralTree::getValue)
-                        .map(String.class::cast);
-            }
-
-            throw new UnsupportedOperationException("Unsupported assignment expression: "
-                    + expression.getClass().getCanonicalName());
-        });
     }
 
     private VisitorStateModifications() {}
